+++ noodles-bam/examples/bam_split.rs
@@ -52,7 +52,7 @@ fn write_headers(writers: &mut Writers, header: &sam::Header) -> io::Result<()> 
 fn find_read_group(data: &sam::record::Data) -> io::Result<Option<String>> {
     use sam::record::data::field::{Tag, Type};
 
-    match data.get(Tag::ReadGroup) {
+    match data.get(&Tag::ReadGroup) {
         Some(value) => value.as_str().map(|s| Some(s.into())).ok_or_else(|| {
             io::Error::new(
                 io::ErrorKind::InvalidData,