+++ noodles-sam/examples/sam_filter_data.rs
@@ -7,7 +7,7 @@ use std::{env, io};
 use noodles_sam::{self as sam, alignment::Record, record::data::field::Tag};
 
 fn is_unique_record(record: &Record) -> io::Result<bool> {
-    match record.data().get(Tag::AlignmentHitCount) {
+    match record.data().get(&Tag::AlignmentHitCount) {
         Some(value) => value.as_int().map(|hits| hits == 1).ok_or_else(|| {
             io::Error::new(
                 io::ErrorKind::InvalidData,